Rename helper to wrapPlugin
